move COCONames from ObjectDetector into new file
