+++ Assets/Scripts/COCONames.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 61a9d1183e3744856abe33e317ddac07
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
